Added exit code to getfile, made startupscript and edited README.md
Automatic processing should be a fact

+++ 1_source/1_raspberry-pi/GetFileFromDirectoryAndSendToServer.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 #saved at /home/pi
+if [ $ACTION == download ];then
 dir='/mnt/ramdisk' #directory where we store our photo's temporary
-file=$(/bin/ls -1 "$dir" | head -1) #list one file (there can only be one file at a time)
+file=$(/bin/ls -1 "$dir" | head -1) #list one file (there can only be one file processed at a time
 path=$(readlink --canonicalize "$dir/$file") #make a path to the file
 server='http://www.verhofstadt.eu/upload.php'
 mac=$(ifconfig eth0 | grep -o -E '([[:xdigit:]]{1,2}:){5}[[:xdigit:]]{1,2}')
 #use this line for uploading, hope it works if bernd has the site online
-#curl -F "userid=1" -F "filecomment=This is an image file" -F "image=@/home/user1/Desktop/test.jpg" localhost/uplo$
+#curl -F "userid=1" -F "filecomment=This is an image file" -F "image=@/home/user/desktop/img.jpg" localhost/upload.php
 curl -F "mac=$mac" -F "image=$path" $server
 rm $path #remove file when it is send, this will remove nothing if there is no file
+else
+        exit 0
+fi
+
